test: move integration tests to a separate directory (#238)
This allows us to focus solely on unit tests, plus, we can tag
integration tests using @moduletag on the module instead of each test
case.
