+++ test/ae_mdw_web/controllers/block_controller_test.exs
@@ -4,18 +4,12 @@ defmodule AeMdwWeb.BlockControllerTest do
   alias :aeser_api_encoder, as: Enc
   alias AeMdw.Validate
   alias AeMdw.Db.{Model, Format}
-  alias AeMdwWeb.{BlockController, TestUtil}
-  alias AeMdwWeb.Continuation, as: Cont
+  alias AeMdwWeb.TestUtil
   alias AeMdw.Error.Input, as: ErrInput
-  alias AeMdw.EtsCache
   require Model
 
   import AeMdw.Db.Util
   import Mock
-
-  @blocks_table AeMdwWeb.BlockController.table()
-  @default_limit 10
-  @blocks_cache_threshold 6
 
   describe "block" do
     test "get key block by hash", %{conn: conn} do
@@ -56,243 +50,6 @@ defmodule AeMdwWeb.BlockControllerTest do
     end
   end
 
-  describe "blocki" do
-    @tag :integration
-    test "get key block by key block index(height)", %{conn: conn} do
-      kbi = 305_200
-      conn = get(conn, "/blocki/#{kbi}")
-
-      assert json_response(conn, 200) == TestUtil.handle_input(fn -> get_blocki(conn.params) end)
-    end
-
-    @tag :integration
-    test "get micro block by given key block index(height) and micro block index", %{conn: conn} do
-      kbi = 305_222
-      mbi = 3
-      conn = get(conn, "/blocki/#{kbi}/#{mbi}")
-
-      assert json_response(conn, 200) == TestUtil.handle_input(fn -> get_blocki(conn.params) end)
-    end
-
-    @tag :integration
-    test "renders error when key block index is invalid", %{conn: conn} do
-      kbi = "invalid"
-      conn = get(conn, "/blocki/#{kbi}")
-
-      assert json_response(conn, 400) ==
-               %{"error" => TestUtil.handle_input(fn -> get_blocki(conn.params) end)}
-    end
-
-    @tag :integration
-    test "renders error when mickro block index is not present", %{conn: conn} do
-      kbi = 305_222
-      mbi = 4999
-      conn = get(conn, "/blocki/#{kbi}/#{mbi}")
-
-      assert json_response(conn, 404) ==
-               %{"error" => TestUtil.handle_input(fn -> get_blocki(conn.params) end)}
-    end
-
-    @tag :integration
-    test "renders error when micro block index is invalid", %{conn: conn} do
-      kbi = 305_222
-      mbi = "invalid"
-      conn = get(conn, "/blocki/#{kbi}/#{mbi}")
-
-      assert json_response(conn, 400) ==
-               %{"error" => TestUtil.handle_input(fn -> get_blocki(conn.params) end)}
-    end
-  end
-
-  describe "blocks" do
-    @tag :integration
-    test "get generations when direction=forward and default limit", %{conn: conn} do
-      direction = "forward"
-      conn = get(conn, "/blocks/#{direction}")
-      response = json_response(conn, 200)
-
-      {:ok, data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, 0},
-          @default_limit
-        )
-
-      assert Enum.count(response["data"]) == @default_limit
-      assert Jason.encode!(response["data"]) == Jason.encode!(data)
-
-      conn_next = get(conn, response["next"])
-      response_next = json_response(conn_next, 200)
-
-      {:ok, next_data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, @default_limit},
-          @default_limit
-        )
-
-      assert Enum.count(response_next["data"]) == @default_limit
-
-      assert Jason.encode!(response_next["data"]) == Jason.encode!(next_data)
-    end
-
-    @tag :integration
-    test "get generations when direction=backward and limit=3", %{conn: conn} do
-      direction = "backward"
-      limit = 3
-      conn = get(conn, "/blocks/#{direction}?limit=#{limit}")
-      response = json_response(conn, 200)
-
-      {:ok, data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, 0},
-          limit
-        )
-
-      assert Enum.count(response["data"]) == limit
-      assert Jason.encode!(response["data"]) == Jason.encode!(data)
-
-      conn_next = get(conn, response["next"])
-      response_next = json_response(conn_next, 200)
-
-      {:ok, next_data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, limit},
-          limit
-        )
-
-      assert Enum.count(response_next["data"]) == limit
-      assert Jason.encode!(response_next["data"]) == Jason.encode!(next_data)
-    end
-
-    @tag :integration
-    test "get generations with numeric range and default limit", %{conn: conn} do
-      range = "305000-305100"
-      conn = get(conn, "/blocks/#{range}")
-      response = json_response(conn, 200)
-
-      {:ok, data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, 0},
-          @default_limit
-        )
-
-      assert Enum.count(response["data"]) == @default_limit
-      assert Jason.encode!(response["data"]) == Jason.encode!(data)
-
-      conn_next = get(conn, response["next"])
-      response_next = json_response(conn_next, 200)
-
-      {:ok, next_data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, @default_limit},
-          @default_limit
-        )
-
-      assert Enum.count(response_next["data"]) == @default_limit
-      assert Jason.encode!(response_next["data"]) == Jason.encode!(next_data)
-    end
-
-    @tag :integration
-    test "get generations with numeric range and limit=1", %{conn: conn} do
-      range = "305000-305100"
-      limit = 1
-      conn = get(conn, "/blocks/#{range}?limit=#{limit}")
-      response = json_response(conn, 200)
-
-      {:ok, data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, 0},
-          limit
-        )
-
-      assert Enum.count(response["data"]) == limit
-      assert Jason.encode!(response["data"]) == Jason.encode!(data)
-
-      conn_next = get(conn, response["next"])
-      response_next = json_response(conn_next, 200)
-
-      {:ok, next_data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, limit},
-          limit
-        )
-
-      assert Enum.count(response_next["data"]) == limit
-      assert Jason.encode!(response_next["data"]) == Jason.encode!(next_data)
-    end
-
-    @tag :integration
-    test "get uncached generations with range", %{conn: conn} do
-      range_begin = last_gen() - @blocks_cache_threshold + 1
-      range_end = last_gen()
-      range = "#{range_begin}-#{range_end}"
-      limit = @blocks_cache_threshold
-      conn = get(conn, "/blocks/#{range}?limit=#{limit}")
-      response = json_response(conn, 200)
-
-      {:ok, data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, 0},
-          limit
-        )
-
-      assert Enum.count(response["data"]) == limit
-      assert Jason.encode!(response["data"]) == Jason.encode!(data)
-      assert nil == EtsCache.get(@blocks_table, range_begin)
-      assert nil == EtsCache.get(@blocks_table, range_end)
-
-      # assert there's nothing next
-      conn_next = get(conn, response["next"])
-      response_next = json_response(conn_next, 200)
-
-      assert Enum.empty?(response_next["data"])
-    end
-
-    @tag :integration
-    test "get a mix of uncached and cached generations with range", %{conn: conn} do
-      remaining = 3
-      range_begin = last_gen() - @blocks_cache_threshold + 1 - remaining
-      range_end = last_gen()
-      range = "#{range_begin}-#{range_end}"
-      limit = @blocks_cache_threshold
-
-      conn = get(conn, "/blocks/#{range}?limit=#{limit}")
-      response = json_response(conn, 200)
-
-      {:ok, data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, 0},
-          limit
-        )
-
-      assert Enum.count(response["data"]) == limit
-      assert Jason.encode!(response["data"]) == Jason.encode!(data)
-      assert {%{"height" => ^range_begin}, _} = EtsCache.get(@blocks_table, range_begin)
-      assert nil == EtsCache.get(@blocks_table, range_end)
-      assert nil == EtsCache.get(@blocks_table, range_end - @blocks_cache_threshold + 1)
-
-      conn_next = get(conn, response["next"])
-      response_next = json_response(conn_next, 200)
-
-      {:ok, next_data, _has_cont?} =
-        Cont.response_data(
-          {BlockController, :blocks, %{}, conn.assigns.scope, limit},
-          limit
-        )
-
-      assert Enum.count(response_next["data"]) == remaining
-      assert Jason.encode!(response_next["data"]) == Jason.encode!(next_data)
-    end
-
-    @tag :integration
-    test "renders error when the range is invalid", %{conn: conn} do
-      range = "invalid"
-      conn = get(conn, "/blocks/#{range}")
-
-      assert json_response(conn, 400) ==
-               %{"error" => "invalid range: #{range}"}
-    end
-  end
-
   ################
 
   defp get_block(enc_block_hash) when is_binary(enc_block_hash) do
@@ -317,14 +74,6 @@ defmodule AeMdwWeb.BlockControllerTest do
       [] ->
         raise ErrInput.NotFound, value: block_index
     end
-  end
-
-  defp get_blocki(%{"kbi" => kbi} = req) do
-    mbi = Map.get(req, "mbi", "-1")
-
-    (kbi <> "/" <> mbi)
-    |> Validate.block_index!()
-    |> get_block()
   end
 
   defp sample_key_block do
