+++ test/integration/ae_mdw_web/controllers/aex9_controller_test.exs
@@ -1,8 +1,9 @@
-defmodule AeMdwWeb.Aex9ControllerTest do
+defmodule Integration.AeMdwWeb.Aex9ControllerTest do
   use AeMdwWeb.ConnCase, async: false
 
+  @moduletag :integration
+
   describe "balance_for_hash" do
-    @tag :integration
     test "gets balance for hash", %{conn: conn} do
       mb_hash = "mh_2NkfQ9p29EQtqL6YQAuLpneTRPxEKspNYLKXeexZ664ZJo7fcw"
       contract_id = "ct_RDRJC5EySx4TcLtGRWYrXfNgyWzEDzssThJYPd9kdLeS5ECaA"
@@ -20,7 +21,6 @@ defmodule AeMdwWeb.Aex9ControllerTest do
   end
 
   describe "balances_for_hash" do
-    @tag :integration
     test "gets balances for hash", %{conn: conn} do
       mb_hash = "mh_2NkfQ9p29EQtqL6YQAuLpneTRPxEKspNYLKXeexZ664ZJo7fcw"
       contract_id = "ct_RDRJC5EySx4TcLtGRWYrXfNgyWzEDzssThJYPd9kdLeS5ECaA"
