+++ test/integration/ae_mdw_web/controllers/contract_controller_test.exs
@@ -1,4 +1,4 @@
-defmodule AeMdwWeb.ContractControllerTest do
+defmodule Integration.AeMdwWeb.ContractControllerTest do
   use AeMdwWeb.ConnCase
   @moduletag :integration
