+++ test/integration/ae_mdw_web/controllers/oracle_controller_test.exs
@@ -1,4 +1,4 @@
-defmodule AeMdwWeb.OracleControllerTest do
+defmodule Integration.AeMdwWeb.OracleControllerTest do
   use AeMdwWeb.ConnCase
 
   alias :aeser_api_encoder, as: Enc
@@ -10,10 +10,11 @@ defmodule AeMdwWeb.OracleControllerTest do
 
   import AeMdwWeb.Util
 
+  @moduletag :integration
+
   @default_limit 10
 
   describe "oracle" do
-    @tag :integration
     test "get oracle information for given oracle id", %{conn: conn} do
       id = "ok_R7cQfVN15F5ek1wBSYaMRjW2XbMRKx7VDQQmbtwxspjZQvmPM"
       conn = get(conn, "/oracle/#{id}")
@@ -26,7 +27,6 @@ defmodule AeMdwWeb.OracleControllerTest do
                end)
     end
 
-    @tag :integration
     test "get oracle information for given oracle id with expand parameter", %{conn: conn} do
       id = "ok_2TASQ4QZv584D2ZP7cZxT6sk1L1UyqbWumnWM4g1azGi1qqcR5"
       conn = get(conn, "/oracle/#{id}?expand")
@@ -40,7 +40,6 @@ defmodule AeMdwWeb.OracleControllerTest do
                end)
     end
 
-    @tag :integration
     test "renders error when oracle id is invalid", %{conn: conn} do
       id = "invalid_oracle_id"
       conn = get(conn, "/oracle/#{id}")
@@ -50,7 +49,6 @@ defmodule AeMdwWeb.OracleControllerTest do
   end
 
   describe "oracles" do
-    @tag :integration
     test "get all oracles with default direction=backward and default limit", %{conn: conn} do
       conn = get(conn, "/oracles")
       response = json_response(conn, 200)
@@ -77,7 +75,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] == next_data
     end
 
-    @tag :integration
     test "get all oracles with direction=forward and limit=3", %{conn: conn} do
       direction = "forward"
       limit = 3
@@ -107,7 +104,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] == next_data
     end
 
-    @tag :integration
     test "get all oracles with limit=7 and expand parameter ", %{conn: conn} do
       limit = 7
       conn = get(conn, "/oracles?limit=#{limit}&expand")
@@ -135,7 +131,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] |> Jason.encode!() == next_data |> Jason.encode!()
     end
 
-    @tag :integration
     test "renders error when direction is invalid", %{conn: conn} do
       direction = "invalid"
       conn = get(conn, "/oracles?direction=#{direction}")
@@ -143,7 +138,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "invalid query: direction=#{direction}"}
     end
 
-    @tag :integration
     test "renders error when limit is invalid", %{conn: conn} do
       limit = "invalid"
       conn = get(conn, "/oracles?limit=#{limit}")
@@ -151,7 +145,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "invalid limit: #{limit}"}
     end
 
-    @tag :integration
     test "renders error when limit is to large", %{conn: conn} do
       limit = 10_000
       conn = get(conn, "/oracles?limit=#{limit}")
@@ -159,7 +152,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "limit too large: #{limit}"}
     end
 
-    @tag :integration
     test "renders error when the access is random ", %{conn: conn} do
       limit = 6
       page = 2
@@ -170,7 +162,6 @@ defmodule AeMdwWeb.OracleControllerTest do
   end
 
   describe "inactive_oracles" do
-    @tag :integration
     test "get inactive oracles with default direction=backward and default limit", %{conn: conn} do
       conn = get(conn, "/oracles/inactive")
       response = json_response(conn, 200)
@@ -197,7 +188,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] == next_data
     end
 
-    @tag :integration
     test "get inactive oracles with direction=forward and limit=5", %{conn: conn} do
       direction = "forward"
       limit = 5
@@ -228,7 +218,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] == next_data
     end
 
-    @tag :integration
     test "get inactive oracles with limit=1 and expand parameter ", %{conn: conn} do
       limit = 1
       conn = get(conn, "/oracles/inactive?limit=#{limit}&expand")
@@ -257,7 +246,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] |> Jason.encode!() == next_data |> Jason.encode!()
     end
 
-    @tag :integration
     test "renders error when direction is invalid", %{conn: conn} do
       direction = "invalid"
       conn = get(conn, "/oracles/inactive?direction=#{direction}")
@@ -265,7 +253,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "invalid query: direction=#{direction}"}
     end
 
-    @tag :integration
     test "renders error when limit is invalid", %{conn: conn} do
       limit = "invalid"
       conn = get(conn, "/oracles/inactive?limit=#{limit}")
@@ -273,7 +260,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "invalid limit: #{limit}"}
     end
 
-    @tag :integration
     test "renders error when limit is to large", %{conn: conn} do
       limit = 10_000
       conn = get(conn, "/oracles/inactive?limit=#{limit}")
@@ -281,7 +267,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "limit too large: #{limit}"}
     end
 
-    @tag :integration
     test "renders error when the access is random ", %{conn: conn} do
       limit = 2
       page = 3
@@ -292,7 +277,6 @@ defmodule AeMdwWeb.OracleControllerTest do
   end
 
   describe "active_oracles" do
-    @tag :integration
     test "get active oracles with default direction=backward and limit=1", %{conn: conn} do
       limit = 1
       conn = get(conn, "/oracles/active?limit=#{limit}")
@@ -320,7 +304,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] == next_data
     end
 
-    @tag :integration
     test "get active oracles with direction=forward and limit=1", %{conn: conn} do
       direction = "forward"
       limit = 1
@@ -351,7 +334,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] == next_data
     end
 
-    @tag :integration
     test "get active oracles with limit=1 and expand parameter ", %{conn: conn} do
       limit = 1
       conn = get(conn, "/oracles/active?limit=#{limit}&expand")
@@ -380,7 +362,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert response_next["data"] |> Jason.encode!() == next_data |> Jason.encode!()
     end
 
-    @tag :integration
     test "renders error when direction is invalid", %{conn: conn} do
       direction = "invalid"
       conn = get(conn, "/oracles/active?direction=#{direction}")
@@ -388,7 +369,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "invalid query: direction=#{direction}"}
     end
 
-    @tag :integration
     test "renders error when limit is invalid", %{conn: conn} do
       limit = "invalid"
       conn = get(conn, "/oracles/active?limit=#{limit}")
@@ -396,7 +376,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "invalid limit: #{limit}"}
     end
 
-    @tag :integration
     test "renders error when limit is to large", %{conn: conn} do
       limit = 10_000
       conn = get(conn, "/oracles/active?limit=#{limit}")
@@ -404,7 +383,6 @@ defmodule AeMdwWeb.OracleControllerTest do
       assert json_response(conn, 400) == %{"error" => "limit too large: #{limit}"}
     end
 
-    @tag :integration
     test "renders error when the access is random ", %{conn: conn} do
       limit = 2
       page = 3
