+++ test/integration/ae_mdw_web/websocket_test.exs
@@ -4,6 +4,8 @@ defmodule AeMdwWeb.WebsocketTest do
   alias AeMdwWeb.Websocket.Listener
   alias Support.WsClient
 
+  @moduletag :integration
+
   setup_all do
     url = "ws://localhost:4001/websocket"
     {:ok, client1} = WsClient.start_link(url)
@@ -174,7 +176,6 @@ defmodule AeMdwWeb.WebsocketTest do
     ]
   end
 
-  @tag :integration
   test "subscribe and unsubscribe to keyblocks, microblocks, transactions and object", setup do
     # subscribe to keyblocks, microblocks, transactions and object
     assert :ok == WsClient.subscribe(setup.client1, :key_blocks)
@@ -242,7 +243,6 @@ defmodule AeMdwWeb.WebsocketTest do
     assert_receive [], 200
   end
 
-  @tag :integration
   test "subscribe to unsupported payload and invalid targets", setup do
     # subscribe to unsupported payload and invalid targets
     assert :ok == WsClient.subscribe(setup.client1, :unsupported_payload)
